+++ requirements.txt
@@ -1,6 +1,5 @@
 allure-pytest==2.13.5
 pydantic==2.9.2
 pytest==8.3.3
-pytest-xdist==3.6.1
 requests==2.32.3
 rstr==3.2.2
