+++ run_test_service.py
@@ -4,6 +4,7 @@ from os.path import join as os_path_join
 
 
 def main():
+    os_system('git clone https://github.com/sun6r0/test-service.git')
     chdir(os_path_join(getcwd(), 'test-service'))
     os_system('docker compose up --build -d')
