+++ tests/fixtures.py
@@ -1,6 +1,8 @@
 from random import randint
+from urllib.parse import urljoin
 
 import pytest
+import requests
 from pydantic import BaseModel
 from rstr import letters
 
@@ -36,3 +38,10 @@ def entity_object():
     def return_entity(input_data: dict):
         return Entity.model_validate(input_data)
     return return_entity
+
+
+@pytest.fixture
+def delete_data():
+    def delete_test_data(input_id: int | str):
+        requests.delete(urljoin('http://localhost:8080/api/delete/', str(input_id)))
+    return delete_test_data
